requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
email-validator==2.1.0
dnspython==2.4.2
Werkzeug==3.0.1
gunicorn==21.2.0